Simplify and fix data types for Byte, Word, and DoubleWord classes
Byte = uint8_t
Word = uint16_t
DoubleWord = uint32_t

This fixes the data type sizes on different platforms

+++ DoubleWord.hpp
@@ -17,27 +17,27 @@ namespace Binary
     class DoubleWord
     {
     public:
-        DoubleWord(unsigned long value = 0) : myValue (value) { };
+        DoubleWord(uint32_t value = 0) : myValue (value) { };
         ~DoubleWord() = default;
         
         template<typename T>
         DoubleWord(const std::initializer_list<T> data)
         {
-            myValue = static_cast<unsigned long>(*data.begin());
+            myValue = static_cast<uint32_t>(*data.begin());
         }
         
         DoubleWord& operator=(DoubleWord const& original);
-        DoubleWord& operator=(unsigned long value);
+        DoubleWord& operator=(uint32_t value);
         
         static DoubleWord readLittleEndian(std::istream& sourceStream);
         static DoubleWord readBigEndian(std::istream& sourceStream);
         void writeLittleEndian(std::ostream& destinationStream) const;
         void writeBigEndian(std::ostream& destinationStream) const;
         
-        operator unsigned long () const {return myValue;}
+        operator uint32_t () const {return myValue;}
         
     private:
-        unsigned long myValue;
+        uint32_t myValue;
     };
 }
